\name{HRVData}
\docType{data}
\alias{HRVData}
\title{HRVData}
\description{
HRVData structure containing the occurrence times of the hearbeats of patient suffering from 
paraplegia and hypertension. The subject from whom the HR was obtained is a patient suffering
from paraplegia and hypertension (systolic blood pressure above 200 mmHg). During
the recording, he is supplied with prostaglandin E1 (a vasodilator that is rarely
employed) and systolic blood pressure fell to 100 mmHg for over an hour. Then, the
blood pressure was slowly recovering until 150 mmHg, more or less
}
\usage{data(HRVData)}
\format{A HRVData structure containing the occurrence times of the heartbeats}
\seealso{\link{HRVProcessedData}}
\keyword{datasets}
