\name{ListEpisodes}
\alias{ListEpisodes}
\title{ Episodes listing }
\description{
  Lists episodes included in a RHRV record
}
\usage{
ListEpisodes(HRVData, TimeHMS = FALSE)
}
\arguments{
  \item{HRVData}{ Data structure that stores the beats register and information related to it }
  \item{TimeHMS}{ Boolean argument to print times in H:M:S format }
}
\references{
L. Rodriguez-Linares, A. Mendez, M. Lado, D. Olivieri, X. Vila, I. Gomez-Conde, "An open source tool for heart rate variability spectral analysis", Computer Methods and Programs in Biomedicine 103, 39-50, doi:10.1016/j.cmpb.2010.05.012 (2011)
}
\author{
M. Lado, A. Mendez, D. Olivieri, L. Rodriguez, X. Vila
}
\keyword{ Episodes }
\keyword{ Tags }
