\name{plot_hr}
\alias{plot_hr}
\title{ Simple plot with configurable size and colour }
\description{
  Plots in a simple way the heart rate signal
}
\usage{
plot_hr(data, width = 12, height = 5, colour = 4)
}
\arguments{
  \item{data}{ heart rate signal to be plotted }
  \item{width}{ width of the plotting window }
  \item{height}{ height of the plotting window }
  \item{colour}{ colour used to plot the signal }
}
\references{
L. Rodriguez-Linares, X. Vila, A. Mendez, M. Lado, D. Olivieri, "RHRV: An R-based software package for heart rate variability analysis of ECG recordings," 3rd Iberian Conference in Systems and Information Technologies (CISTI 2008), 21-23 June 2008.
}
\author{
M. Lado, A. Mendez, D. Olivieri, L. Rodriguez, X. Vila
}
\keyword{ aplot }
