\name{pw_from_file}
\alias{pw_from_file}
\title{ Obtains power per band from a file }
\description{
  A function to illustrate the calculation of the power per band from a file in ascii format containing the data of beat positions
}
\usage{
pw_from_file(file, size, shift, fsamp = 4)
}
\arguments{
  \item{file}{ the file to be analysed }
  \item{size}{ size of window (sec.) }
  \item{shift}{ displacement of window (sec.) }
  \item{fsamp}{ sampling frequency }
}
\references{
L. Rodriguez-Linares, X. Vila, A. Mendez, M. Lado, D. Olivieri, "RHRV: An R-based software package for heart rate variability analysis of ECG recordings," 3rd Iberian Conference in Systems and Information Technologies (CISTI 2008), 21-23 June 2008.
}
\author{
M. Lado, A. Mendez, D. Olivieri, L. Rodriguez, X. Vila
}
\keyword{ misc }

