% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHRVEasy.R
\name{SaveHRVIndices}
\alias{SaveHRVIndices}
\title{Save the HRV indices as an excel file}
\usage{
SaveHRVIndices(
  RHRVEasyResultObject,
  saveHRVIndicesInPath = ".",
  filename = NULL
)
}
\arguments{
\item{RHRVEasyResultObject}{An object of class `RHRVEasyResult` as returned
by `RHRVEasy()`}

\item{saveHRVIndicesInPath}{The path where the HRV indices will be saved as
an excel file. The name of the file is automatically created based on the
groups being compared.}

\item{filename}{Filename of the excel file. If not provided, the name of the file
is built using the names of the groups being compared.}
}
\description{
Save the HRV indices as an excel file
}
